requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
python-jose[cryptography]
requests
pydantic
passlib[bcrypt]